+++ docs/index.html
@@ -1213,7 +1213,7 @@
     <script>
       // Customize
       eclipseFdnVideos.replace({
-        selector: ".eclipsefdn-videoss",
+        selector: ".eclipsefdn-video",
         resolution: "16by9",
         cookie: {
           name: "eclipse_cookieconsent_status",
