+++ js/main.js
@@ -1,7 +1,15 @@
-import jquery from 'jquery';
-import bootstrap from "bootstrap";
-import solstice from "./src/eclipsefdn.solstice";
-import eclipseFdnAdopters  from './src/api/eclipsefdn.adopters'
 
-window.eclipseFdnAdopters = eclipseFdnAdopters;
+/*!
+ * Copyright (c) 2021 Eclipse Foundation, Inc.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * Contributors:
+ *   Christopher Guindon <chris.guindon@eclipse-foundation.org>
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
 
+import solstice from "./src";
